LPS-111957 Migrates JournalArticleSelector to use React only and applies some naming standards
