+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/index.js
@@ -12,5 +12,7 @@
  * details.
  */
 
+export {connectStore} from './util/connectStore.es';
+export {FieldBaseProxy} from './FieldBase/ReactFieldBase.es';
 export {default as FieldBase} from './FieldBase/FieldBase.es';
 export {default as getConnectedReactComponentAdapter} from './util/ReactComponentAdapter.es';
