+++ modules/apps/journal/journal-article-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/JournalArticleSelector.js
@@ -12,173 +12,133 @@
  * details.
  */
 
-import 'dynamic-data-mapping-form-field-type';
-import Component from 'metal-component';
-import Soy from 'metal-soy';
-import {Config} from 'metal-state';
+import ClayButton from '@clayui/button';
+import ClayForm, {ClayInput} from '@clayui/form';
+import {
+	FieldBaseProxy,
+	connectStore,
+	getConnectedReactComponentAdapter,
+} from 'dynamic-data-mapping-form-field-type';
+import {ItemSelectorDialog} from 'frontend-js-web';
+import React, {useEffect, useState} from 'react';
 
-import './JournalArticleSelectorAdapter.soy';
+const JournalArticleSelector = ({
+	disabled,
+	inputValue,
+	itemSelectorURL,
+	name,
+	onChange,
+	portletNamespace,
+}) => {
+	const [article, setArticle] = useState({});
 
-import './JournalArticleSelectorRegister.soy';
-
-import './ReactJournalArticleSelectorAdapter';
-import templates from './JournalArticleSelector.soy';
-
-class JournalArticleSelector extends Component {
-	dispatchEvent(event, name, value) {
-		this.emit(name, {
-			fieldInstance: this,
-			originalEvent: event,
-			value,
+	useEffect(() => {
+		setArticle({
+			...JSON.parse(inputValue || '{}'),
 		});
-	}
+	}, [inputValue]);
 
-	_handleOnDispatch(event) {
-		switch (event.type) {
-			case 'value':
-				this.dispatchEvent(event, 'fieldEdited', event.payload);
-				break;
-			case 'blur':
-				this.dispatchEvent(
-					event.payload,
-					'fieldBlurred',
-					event.payload.target.value
-				);
-				break;
-			case 'focus':
-				this.dispatchEvent(
-					event.payload,
-					'fieldFocused',
-					event.payload.target.value
-				);
-				break;
-			default:
-				console.error(new TypeError(`There is no type ${event.type}`));
-				break;
+	const handleClearClick = () => {
+		setArticle({});
+		onChange('');
+	};
+
+	const handleFieldChanged = event => {
+		const selectedItem = event.selectedItem;
+
+		if (selectedItem && selectedItem.value) {
+			setArticle(selectedItem.value);
+			onChange(selectedItem.value);
 		}
-	}
-}
+	};
 
-JournalArticleSelector.STATE = {
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
+	const handleItemSelectorTriggerClick = event => {
+		event.preventDefault();
 
-	errorMessage: Config.string(),
+		const itemSelectorDialog = new ItemSelectorDialog({
+			eventName: `${portletNamespace}selectJournalArticle`,
+			singleSelect: true,
+			title: Liferay.Language.get('journal-article'),
+			url: itemSelectorURL,
+		});
 
-	/**
-	 * @default false
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?bool}
-	 */
+		itemSelectorDialog.on('selectedItemChange', handleFieldChanged);
 
-	evaluable: Config.bool().value(false),
+		itemSelectorDialog.open();
+	};
 
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
+	return (
+		<ClayForm.Group style={{marginBottom: '0.5rem'}}>
+			<ClayInput.Group>
+				<ClayInput.GroupItem className="d-none d-sm-block" prepend>
+					<input name={name} type="hidden" value={inputValue} />
 
-	fieldName: Config.string(),
+					<ClayInput
+						className="bg-light"
+						disabled={disabled}
+						onClick={handleItemSelectorTriggerClick}
+						readOnly
+						type="text"
+						value={article.title || ''}
+					/>
+				</ClayInput.GroupItem>
 
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
+				<ClayInput.GroupItem append shrink>
+					<ClayButton
+						disabled={disabled}
+						displayType="secondary"
+						onClick={handleItemSelectorTriggerClick}
+						type="button"
+					>
+						{Liferay.Language.get('select')}
+					</ClayButton>
+				</ClayInput.GroupItem>
 
-	label: Config.string(),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	name: Config.string().required(),
-
-	/**
-	 * @default '000000'
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	predefinedValue: Config.string().value('000000'),
-
-	/**
-	 * @default false
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?bool}
-	 */
-
-	readOnly: Config.bool().value(false),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(bool|undefined)}
-	 */
-
-	repeatable: Config.bool().value(false),
-
-	/**
-	 * @default false
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(bool|undefined)}
-	 */
-
-	required: Config.bool().value(false),
-
-	/**
-	 * @default true
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(bool|undefined)}
-	 */
-
-	showLabel: Config.bool().value(true),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	spritemap: Config.string(),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	tip: Config.string(),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof JournalArticleSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	value: Config.string(),
+				{article.classPK && (
+					<ClayInput.GroupItem shrink>
+						<ClayButton
+							disabled={disabled}
+							displayType="secondary"
+							onClick={handleClearClick}
+							type="button"
+						>
+							{Liferay.Language.get('clear')}
+						</ClayButton>
+					</ClayInput.GroupItem>
+				)}
+			</ClayInput.Group>
+		</ClayForm.Group>
+	);
 };
 
-Soy.register(JournalArticleSelector, templates);
+const JournalArticleSelectorProxy = connectStore(
+	({
+		emit,
+		itemSelectorURL,
+		name,
+		portletNamespace,
+		predefinedValue,
+		readOnly,
+		value,
+		...otherProps
+	}) => (
+		<FieldBaseProxy {...otherProps} name={name} readOnly={readOnly}>
+			<JournalArticleSelector
+				disabled={readOnly}
+				inputValue={value && value !== '' ? value : predefinedValue}
+				itemSelectorURL={itemSelectorURL}
+				name={name}
+				onChange={value => emit('fieldEdited', {}, value)}
+				portletNamespace={portletNamespace}
+			/>
+		</FieldBaseProxy>
+	)
+);
 
-export {JournalArticleSelector};
-export default JournalArticleSelector;
+const ReactJournalArticleSelectorAdapter = getConnectedReactComponentAdapter(
+	JournalArticleSelectorProxy,
+	'journal_article'
+);
+
+export {ReactJournalArticleSelectorAdapter};
+export default ReactJournalArticleSelectorAdapter;
