--- modules/apps/journal/journal-article-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/JournalArticleSelector.soy
@@ -1,55 +1,0 @@
-{namespace JournalArticleSelector}
-
-{template .render}
-	{@param name: string}
-	{@param? _handleOnDispatch: any}
-	{@param? displayErrors: bool}
-	{@param? errorMessage: string}
-	{@param? id: string}
-	{@param? itemSelectorURL: string}
-	{@param? label: string}
-	{@param? predefinedValue: string}
-	{@param? portletNamespace: string}
-	{@param? readOnly: bool}
-	{@param? repeatable: bool}
-	{@param? required: bool}
-	{@param? showLabel: bool}
-	{@param? spritemap: string}
-	{@param? tip: string}
-	{@param? valid: bool}
-	{@param? value: string}
-	{@param? visible: bool}
-
-	{let $inputValue: ($value and $value != '') ? $value : $predefinedValue /}
-
-	{call FieldBase.render}
-		{param contentRenderer kind="html"}
-			{call JournalArticleSelectorAdapter.render}
-				{param events: ['dispatch': $_handleOnDispatch] /}
-				{param ref: $name /}
-				{param store: [
-					'inputValue': $inputValue,
-					'itemSelectorURL': $itemSelectorURL,
-					'name': $name,
-					'readOnly': $readOnly,
-					'portletNamespace': $portletNamespace,
-					'spritemap': $spritemap
-				] /}
-			{/call}
-		{/param}
-		{param displayErrors: $displayErrors /}
-		{param errorMessage: $errorMessage /}
-		{param id: $id /}
-		{param label: $label /}
-		{param name: $name /}
-		{param readOnly: $readOnly /}
-		{param repeatable: $repeatable /}
-		{param required: $required /}
-		{param showLabel: $showLabel /}
-		{param spritemap: $spritemap /}
-		{param tip: $tip /}
-		{param valid: $valid /}
-		{param ref: 'fieldBase' /}
-		{param visible: $visible /}
-	{/call}
-{/template}
