--- modules/apps/journal/journal-article-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/JournalArticleSelectorAdapter.soy
@@ -1,8 +1,0 @@
-{namespace JournalArticleSelectorAdapter}
-
-/**
- * JournalArticleAdapter
- */
-{template .render}
-	<div key="ddm-field-type-journal-article-selector"></div>
-{/template}
