--- modules/apps/journal/journal-article-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/JournalArticleSelectorRegister.soy
@@ -1,5 +1,0 @@
-{namespace JournalArticleSelectorRegister}
-
-{deltemplate PageRenderer.RegisterFieldType variant="'journal_article'"}
-	{call JournalArticleSelector.render data="all" /}
-{/deltemplate}
